server-systemd: Migrate journal code into a package
This moves server code into a package. Put the main journal
into a component that is shown in an iframe.

The shell loads the server AMD module somewhat awkwardly, but
that's because it's not AMD code.

Fixes #1418
Closes #1473
