+++ pkg/base/cockpit.css
@@ -444,3 +444,29 @@ code, kbd, pre, samp {
 .cockpit-modal-sm {
     width: 300px;
 }
+
+/* Extra content header */
+
+.content-header-extra {
+    background: #f5f5f5;
+    border-bottom: 1px solid #ddd;
+    padding: 10px 20px;
+    width: 100%;
+    position: fixed;
+    z-index: 900;
+    top: 0;
+}
+
+.content-header-extra .btn-group:not(:first-child) {
+    padding-left: 20px;
+}
+
+/* Server stuff, not enough for own CSS yet */
+
+.journal-lines .cockpit-logline:last-child {
+    border-bottom: 1px solid #ddd;
+}
+
+.journal-lines .cockpit-logline:hover {
+    cursor: pointer;
+}
