+++ pkg/shell/Makefile.am
@@ -20,8 +20,6 @@ shell_js_PARTS = \
 	pkg/shell/cockpit-realms.js \
 	pkg/shell/cockpit-shutdown.js \
 	pkg/shell/cockpit-services.js \
-	pkg/shell/cockpit-journal.js \
-	pkg/shell/cockpit-journal-renderer.js \
 	pkg/shell/cockpit-accounts.js \
 	pkg/shell/cockpit-docker.js \
 	$(NULL)
@@ -77,7 +75,6 @@ shellimages_DATA = \
 CHECK_HTML = \
 	pkg/shell/test-dummy.html \
 	pkg/shell/test-util.html \
-	pkg/shell/test-journal-renderer.html \
 	pkg/shell/test-rest.html \
 	pkg/shell/test-dbusx.html \
 	$(NULL)
