+++ pkg/shell/shell.css
@@ -149,16 +149,16 @@ html, body {
 
 /* Extra content header */
 
-#content-header-extra {
+#shell-header-extra {
     background: #f5f5f5;
     border-bottom: 1px solid #ddd;
 }
 
-#content-header-extra .btn-group {
+#shell-header-extra .btn-group {
     margin: 10px 0;
 }
 
-#content-header-extra .btn-group:not(:first-child) {
+#shell-header-extra .btn-group:not(:first-child) {
     padding-left: 20px;
 }
 
@@ -212,14 +212,6 @@ ul.available-interfaces-group {
 
 #display-language-list option {
     padding: 10px;
-}
-
-#journal-box .cockpit-logline:last-child {
-    border-bottom: 1px solid #ddd;
-}
-
-#journal-box .cockpit-logline:hover {
-    cursor: pointer;
 }
 
 .server-graph:hover {
