+++ po/POTFILES.in
@@ -1,11 +1,12 @@
 [encoding: UTF-8]
 # List of source files containing translatable strings.
 # Please keep this file sorted alphabetically.
+pkg/server-systemd/log.html
+pkg/server-systemd/server.js
 [type: gettext/glade] pkg/shell/shell.html
 pkg/shell/cockpit-about.js
 pkg/shell/cockpit-cpu-status.js
 pkg/shell/cockpit-dashboard.js
-pkg/shell/cockpit-journal.js
 pkg/shell/cockpit-language.js
 pkg/shell/cockpit-main.js
 pkg/shell/cockpit-memory-status.js
